fix: type이 제대로 불러와지지 않던 문제 해결
