+++ src/types/type.d.ts
@@ -1,17 +1,3 @@
-import { COMPANY_TABLE } from '../constants/table';
-
-interface PaymentsInputFieldProps {
-  className?: string;
-  placeholder?: string;
-  maxLength?: number;
-  hasError?: boolean;
-  value?: string;
-  handleValueChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  handleOnFocus?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  handleOnBlur?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  width?: number;
-}
-
 type InputStates = InputState[];
 
 interface InputState {
